Add files via upload
Step 1: Now, Initialize the node package by typing in the following command in the terminal:
npm init -y

Step 2: This will create a package.json file for us so that we can install the necessary libraries. Now type in the command:
npm install express socket.io

Steps to run the application: Write the below code in the terminal to run the server:
node index.js

Output: Open your browser and open two tabs with the URL: http://localhost:3000 so that we have two clients that can chat with each other. Following is the output of our chat app: